refactoring : profile (#51)
* refactoring : created store directory and replaced profileApi into.

* refactoring : refactored LogoutButton.

* refactoring : refactored data fetching hooks.

* refactoring : refactored hook import in component.
